check is running before kill

+++ images/03.ServiceImages/email/home/engines/scripts/signal/_kill_postfix.sh
@@ -3,6 +3,10 @@
 
  if  test -f /var/spool/postfix/pid/master.pid
   then
-    kill $1 `cat /var/spool/postfix/pid/master.pid`
+   kill -0 `cat /var/spool/postfix/pid/master.pid` 
+    if test $? -eq 0
+     then
+    	kill $1 `cat /var/spool/postfix/pid/master.pid`
+     fi
  fi
